+++ Plugin.cs
@@ -1,17 +1,13 @@
 ﻿using BepInEx;
 using BepInEx.Bootstrap;
-using BepInEx.Configuration;
 using BepInEx.Logging;
 using LethalCompTime.Patches;
 using LethalCompTime.Configs;
 using HarmonyLib;
-using LobbyCompatibility.Enums;
-using LobbyCompatibility.Attributes;
 
 namespace LethalCompTime
 {
     [BepInPlugin("com.github.cdusold.LethalCompTime", PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    [LobbyCompatibility(CompatibilityLevel.ClientOnly, VersionStrictness.None)]
     public class Plugin : BaseUnityPlugin
     {
         private readonly Harmony harmony = new Harmony("cdusold.LethalCompTime");